requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.15.6
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bandit==1.8.3
beautifulsoup4==4.13.3
bitarray==3.0.0
black==23.11.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
colorama==0.4.6
comm==0.2.1
config==0.5.1
cryptography==41.0.5
customtkinter==5.2.1
cytoolz==1.0.1
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.7
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.3.0
django-environ==0.12.0
django-extensions==3.2.3
django-rest-passwordreset==1.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
entrypoints==0.4
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==1.0.1
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
executing==2.0.1
Faker==18.11.1
filelock==3.18.0
flake8==7.1.2
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.3.3
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gunicorn==20.1.0
hexbytes==0.3.1
httplib2==0.22.0
identify==2.6.9
idna==3.10
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.14.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lazy-object-proxy==1.9.0
lru-dict==1.3.0
lxml==5.2.1
Markdown==3.10.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.8.8
nodeenv==1.9.1
numpy==2.2.2
O365==2.0.38
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
packaging==23.2
parsimonious==0.10.0
parso==0.8.3
pathspec==0.12.1
pbr==6.1.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
pipenv==2023.6.18
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
prompt-toolkit==3.0.38
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
PyDrive==1.3.1
PyDrive2==1.21.3
pyflakes==3.2.0
pygame==2.5.2
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyJWT==2.10.0
pylint==2.17.5
pymarc==5.2.3
PyMuPDF==1.25.2
pyOpenSSL==23.3.0
pyparsing==3.2.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.0
python-tss-sdk==1.2.3
python_tsp==0.5.0
pytz==2023.3.post1
pyunormalize==16.0.0
PyYAML==6.0.2
pyzmq==25.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rlp==4.0.1
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
stevedore==5.4.1
tabulate==0.8.10
tomli==2.2.1
tomlkit==0.12.1
toolz==1.0.0
tornado==6.3.2
tqdm==4.67.1
traitlets==5.9.0
tsplib95==0.7.1
ttkwidgets==0.13.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
urllib3==2.3.0
user-agents==2.2.0
virtualenv==20.29.3
virtualenv-clone==0.5.7
wcwidth==0.2.6
web3==6.11.1
websockets==14.2
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.15.0
yarl==1.9.2
zeep==4.2.1
reportlab==4.4.10
django-watson==1.6.3
